+++ lib/src/astar_grid.dart
@@ -1,4 +1,5 @@
 import 'dart:async';
+import 'dart:collection';
 import 'dart:math';
 
 import 'package:meta/meta.dart';
@@ -75,33 +76,45 @@ abstract class AstarGrid {
   ///
   /// return null if path not found, no way on labyrinth
   @internal
-  ANode? getWinner(ANode current, ANode end, {int ceilSize = 4}) {
-    final waitList = <ANode>[];
+  ANode? getWinner(ANode current, ANode end) {
+    final splay = SplayTreeMap<int, List<ANode>>();
     if (end == current) return current;
     final neighbors = current.neighbors;
     for (var n in neighbors) {
       n.visited = true;
       n.parent = current;
       n.g = n.weight;
-      waitList.add(n);
+      splay.update(
+        n.f,
+        (value) {
+          return [...value, n];
+        },
+        ifAbsent: () {
+          return [n];
+        },
+      );
     }
 
-    waitList.sort((a, b) => b.f.compareTo(a.f));
-    int maxLength = min(waitList.length, ceilSize);
-    while (waitList.isNotEmpty) {
-      /// iteration n nodes per time
-      for (var i = 0; i < maxLength; i++) {
-        final c = waitList.removeLast();
+    while (splay.isNotEmpty) {
+      final key = splay.firstKey();
+      final list = splay.remove(key);
+      for (var c in list!) {
         if (end == c) return c;
         for (var n in c.neighbors) {
           if (n.visited) continue;
           analyzeDistance(n, end, parent: c);
           n.visited = true;
-          waitList.add(n);
+          splay.update(
+            n.f,
+            (value) {
+              return [...value, n];
+            },
+            ifAbsent: () {
+              return [n];
+            },
+          );
         }
       }
-      maxLength = min(waitList.length, ceilSize);
-      waitList.sort((a, b) => b.f.compareTo(a.f));
     }
     return null;
   }
@@ -118,32 +131,6 @@ abstract class AstarGrid {
       path.add(nextNode);
     }
 
-    return path;
-  }
-
-  List<ANode> reconstructNormalized(Point<int> end) {
-    ANode astar = grid[end.x][end.y];
-    final path = [astar];
-    var nextList = <ANode>[];
-    nextList.addAll(astar.neighbors);
-    if (nextList.isEmpty) return path;
-    nextList.sort((a, b) => b.g.compareTo(a.g));
-    int g = astar.g;
-    int index = 0;
-    int length = nextList.length;
-    while (index < length) {
-      final n = nextList[index];
-      if (!n.visited) continue;
-      if (n.g < g) {
-        g = n.g;
-        path.add(n);
-        index = 0;
-        length = n.neighbors.length;
-        nextList = n.neighbors;
-        nextList.sort((a, b) => b.g.compareTo(a.g));
-      }
-      index++;
-    }
     return path;
   }